Add usage flags to each registered command.
These flags indicate how each command can be used such as only with
websockets or wallet and if it's a notification.  It also exposes an API
to query the usage flags for a method.

Finally, expose a new API call to retrieve a sorted list of all registered
command methods.